requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
bitarray==2.7.3
certifi==2022.12.7
charset-normalizer==3.1.0
cytoolz==0.12.1
Django==4.2
djangorestframework==3.14.0
eth-abi==4.0.0
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
frozenlist==1.3.3
hexbytes==0.3.0
idna==3.4
jsonschema==4.17.3
lru-dict==1.1.8
multidict==6.0.4
parsimonious==0.9.0
protobuf==4.22.1
pycryptodome==3.17
pyrsistent==0.19.3
pytz==2023.3
pywin32==306
regex==2023.3.23
requests==2.28.2
rlp==3.0.0
sqlparse==0.4.3
toolz==0.12.0
tzdata==2023.3
urllib3==1.26.15
web3==6.1.0
websockets==11.0.1
yarl==1.8.2
